Documentation corrections
Changes in both evil-ediff.el and README.org:

Description:

Added a missing word "to" and made a slight rewrite of the description:

Before:
Make ediff a little evil. This configures ediff to be a little more friendly
users of vim-like keybindings. Consult the help buffer (=?=) for more info.

After:
Make ediff a little evil. This configures ediff to be friendlier to users
of vim-like keybindings. Consult the help buffer (=?=) for more info.

Correct the "table describing the bindings":

- Moved down `ediff-jump-to-difference` two lines so that it's grouped with
the other two jump commands.

- Reordered the `ediff-previous-difference` and `ediff-next-difference`
commands so that next is first, because it's probably the most used command,
that also makes them match the order of the scroll commands, down (next)
before up (previous).

- Moved `ediff-suspend` to the bottom, because it's the last command in the
ediff control panel, and alphabetically "z" probably fits better at the bottom.
And moving it also groups the scroll commands together.

- Moved up the "copy A region to B's region" and "copy B region to A's region"
commands below the "navigate to a difference" commands, because they are
probably the second most used commands, one first navigates to a difference and
then copies one set of changes.
  - Rewrote the names without "'s".
  - Added "a" and "b" as keys in the Evil-ediff column, because they still work
  to copy regions, and they are listed in README.org

evil-ediff.el specific changes:

- Removed a duplicate line
;; | ediff-next-difference       | n,SPC            | C-j,n,SPC   |

- Reordered "copy A region to B's region" and "copy B region to A's region" to
sort them alphabetically, that also makes them match the order in README.org.

- Added the toggle highlight command, it was only listed in README.org
;; | toggle highlighting         | h                | H           |

- Separated the original bindings "rarb", and added the missing 3rd binding "rc"
before:
;; | restore old diff            | rarb             | u           |

after:
;; | restore old diff            | ra,rb,rc         | u           |

- The docstring for `defvar evil-ediff-bindings` was missing a space between
"A" and "list":
before:
  "Alist of bindings changed/added in evil-ediff.")

after:
  "A list of bindings changed/added in evil-ediff.")

README.org specific changes:

The Evil-ediff bindings "=l=" and "=h=" for the "copy A region to B's region"
commands, are showing the "=" equal signs around the letters. By looking at
how the other bindings are written in README.org one can see that the inner
equal signs should be removed:
before:
| copy A region to B's region | =a=              | =a=,=l=        |
| copy B region to A's region | =b=              | =b=,=h=        |

after:
| copy A region to B's region | =a=              | =a,l=          |
| copy B region to A's region | =b=              | =b,h=          |